requirements.txt
jupyterlab
nltk
ufal.udpipe
unidecode

